Python 3.6 compatility (e.g. for using it in Windows with TF)
After getting Tensorzoom working in a Linux VM I wanted to run it in a Windows 10 environment.
As Tensorflow for Windows is only available for Python 3 I made the necessary changes.
Tested with: Windows10x64, Python 3.6.5, Tensorflow 1.11.0, Numpy 1.14.5, scikit-image 0.14.0

Most changes were changing the print "" to print("").
When np.load() there has to be added an encoding param.